+++ src/raw/response.rs
@@ -4,6 +4,8 @@
 //! information. See the documentation of the [`RateLimit`] type for more
 //! details.
 
+#![allow(missing_docs)]
+
 use std::collections::HashMap;
 
 use attohttpc::Response;
@@ -252,15 +254,14 @@ pub struct UserPlayingNowPayload {
     pub count: u8,
     pub user_id: String,
     pub listens: Vec<UserPlayingNowListen>,
+    pub playing_now: bool,
 }
 
 /// Type of the [`UserPlayingNowPayload::listens`] field.
 #[derive(Debug, Deserialize)]
 pub struct UserPlayingNowListen {
-    pub user_name: String,
-    pub inserted_at: String,
-    pub recording_msid: String,
     pub track_metadata: UserPlayingNowTrackMetadata,
+    pub playing_now: bool,
 }
 
 /// Type of the [`UserPlayingNowListen::track_metadata`] field.
@@ -318,7 +319,7 @@ pub struct UserListensPayload {
 #[derive(Debug, Deserialize)]
 pub struct UserListensListen {
     pub user_name: String,
-    pub inserted_at: String,
+    pub inserted_at: i64,
     pub listened_at: i64,
     pub recording_msid: String,
     pub track_metadata: UserListensTrackMetadata,
@@ -409,7 +410,7 @@ response_type! {
 /// Type of the [`StatsSitewideArtistsResponse::payload`] field.
 #[derive(Debug, Deserialize)]
 pub struct StatsSitewideArtistsPayload {
-    pub time_ranges: Vec<StatsSitewideArtistsTimeRange>,
+    pub artists: Vec<StatsSitewideArtistsArtist>,
     pub offset: u64,
     pub count: u64,
     pub range: String,
@@ -418,20 +419,10 @@ pub struct StatsSitewideArtistsPayload {
     pub to_ts: i64,
 }
 
-/// Type of the [`StatsSitewideArtistsPayload::time_ranges`] field.
-#[derive(Debug, Deserialize)]
-pub struct StatsSitewideArtistsTimeRange {
-    pub time_range: String,
-    pub artists: Vec<StatsSitewideArtistsArtist>,
-    pub from_ts: i64,
-    pub to_ts: i64,
-}
-
 /// Type of the [`StatsSitewideArtistsTimeRange::artists`] field.
 #[derive(Debug, Deserialize)]
 pub struct StatsSitewideArtistsArtist {
     pub artist_mbids: Option<Vec<String>>,
-    pub artist_msid: String,
     pub artist_name: String,
     pub listen_count: u64,
 }